refactoring game class out of server.js added game class to game-manager
