+++ server/server.js
@@ -1,8 +1,8 @@
-var systems = require('./systems.js');
-var physObjects = require('./physics-objects.js');
+
 
 const PORT = 55000;
-
+var gameManager = require('./game-manager.js');
+gameManager.createManager();
 var server = require('http').createServer();
 global.io = require('socket.io')(server);
 io = global.io;
@@ -81,7 +81,7 @@ io.on('connection', function(client) {
                             game.destroy();
                         }
                         // creating lobbies and multiple games should be easy now the game is handling where the updates go!
-                        game = new Game(lobby.members, 'lobby');
+                        game = gameManager.createNewGame(lobby.members);
                         io.sockets.in('lobby').emit("loadgame");
                         lobby = {};
                         lobby.members = {};
@@ -140,274 +140,5 @@ function getAllPlayers(){
         var player = game.players[socketID];//io.sockets.connected[socketID].player;
         if(player) players.push(player);
     });
-    // Object.keys(io.sockets.connected).forEach(function(socketID){
-    //     console.log(io.sockets.connected[socketID].player);
-    //     var player = io.sockets.connected[socketID].player;
-    //     if(player) players.push(player);
-    // });
     return players;
 }
-
-function randomInt(low, high) {
-    return Math.floor(Math.random() * (high - low) + low);
-}
-function randomVelocity(low, high){
-    let velY = randomInt(low, high);
-    let velX = randomInt(low, high);
-    return{x: velX, y: velY};
-}
-
-
-function isVelocityBetweenMinAndMax(vel, min, max){
-    let absVeloX = Math.abs(vel.x);
-    let absVeloY = Math.abs(vel.y);
-    if(absVeloX + absVeloY < min || absVeloX + absVeloY > max ){
-        return false
-    }else{
-        return true
-    }
-}
-
-
-
-class Game {
-    constructor(membersList, gameid){
-        this.players = {};
-        this.goals = {};
-        this.balls = {};
-        this.gameid = gameid;
-        this.collisions = [];
-        this.collisionManager = new systems.CollisionManager();
-        this.createPlayers(membersList);
-        this.createBall();
-        this.updaterID = systems.addToUpdate(this);
-    }
-
-    createPlayers(membersList){
-        for(let memberkey in membersList){
-            let member = membersList[memberkey];
-            this.createPlayer(member);
-        }
-    }
-
-    createPlayer(member){
-        let startVectors = this.getStartVectors(member.position);
-        let xPos = startVectors.x;
-        let yPos = startVectors.y;
-        let isRotated = this.getIsRotated(member.position)
-        let newPlayer = new physObjects.Player(member.position, xPos , yPos, isRotated, member.character, member.socketid);
-        this.players[member.socketid] = newPlayer;
-        this.createGoal(newPlayer, xPos, yPos, isRotated);
-    }
-
-    createGoal(player, x, y, isRotated){
-        let goalWidth = 1000;
-        let goalHeight = 20;
-        let newGoal = new physObjects.PlayerGoal(x, y, goalWidth, goalHeight, isRotated, player);
-        this.goals[player.id] = newGoal;
-    }
-
-    createBall(){
-        this.lastBallID++;
-        let ballWidth = 48;
-        let newBall = new physObjects.Ball(
-            physObjects.gameHeight/2,
-            physObjects.gameWidth/2,
-            ballWidth/2,
-            true,
-            (ball, bounds)=> { this.onCollisionBallBounds(ball, bounds)
-            });
-
-        this.setBallVelocityBetween(newBall, 5, 5, -10, 10);
-        this.balls[this.lastBallID] = newBall;
-        this.addBallCollisions(newBall);
-
-    }
-
-    addBallCollisions(ball){
-        for(let playerKey in this.players) {
-            let player = this.players[playerKey];
-            this.collisionManager.addCollision(player, ball, () => { this.onCollisionPlayerBall(player, ball) });
-        }
-        for(let goalKey in this.goals) {
-            let goal = this.goals[goalKey];
-            this.collisionManager.addCollision(goal, ball, () => { this.onCollisionGoalBall(goal, ball);});
-        }
-    }
-
-    update(){
-        this.updatePlayerPositions();
-        this.updateBallPositions();
-    }
-
-    updateBallPositions(){
-        for(let ballKey in this.balls) {
-            let ball = this.balls[ballKey];
-            ball.update();
-            //long term we dont need global update
-            global.io.sockets.in(this.gameid).emit('moveball', ball);
-        }
-    }
-
-    updatePlayerPositions(){
-        for(let playerKey in this.players) {
-            let player = this.players[playerKey];
-            player.update();
-            //long term we dont need global update
-            global.io.sockets.in(this.gameid).emit('move', player);
-        }
-    }
-
-    getIsRotated(playerNumber){
-        return !(playerNumber % 2);
-    }
-
-    // get where the player should start
-    getStartVectors(playerNumber){
-        const width = physObjects.gameWidth;
-        const height = physObjects.gameHeight;
-        const playerOffset = 50;
-        switch (playerNumber) {
-            case 0:   return {x: 0 + playerOffset, y: height/2}
-            case 1:   return {x: width/2, y: height - playerOffset}
-            case 2:   return {x: width - playerOffset, y: height/2}
-            case 3:   return {x: width/2, y: 0 + playerOffset}
-        }
-    }
-
-    killPlayer(player){
-        player.isActive = false;
-        // this will leave a collision reference on the collision manager
-        // deleting the reference in this array will prevent any balls created after from adding collision and prevent movemnt updates
-        // a ball adding a collision would not work after the player has been disables
-        // we could potentially change the collision managers to return an ID when an object is added
-        // this will allow us to remove objects completely
-        delete this.players[player.socketid];
-        delete this.goals[player.id];
-        global.io.sockets.in(this.gameid).emit('playerdeath', {id:player.id});
-    }
-
-    onCollisionPlayerBall(player, ball) {
-        // emit this shit for game to know where collision occurred, well, where the ball was when it did
-        global.io.sockets.in(this.gameid).emit('collisionplayer', {player: player, ball: ball});
-
-        let playerWidth = player.width / 2;
-        let playerHeight = player.height / 2;
-        let ballWidth = ball.width / 2;
-        let ballHeight = ball.height / 2;
-
-        let bound;
-        if (ball.x > player.x + playerWidth) {
-            bound = 0; // east of player - bounce right
-            ball.x += ballWidth;
-        } else if (ball.x < player.x - playerWidth) {
-            bound = 2; // west of player - bounce left
-            ball.x -= ballWidth;
-        }
-        
-        if (ball.y > player.y + playerHeight) {
-            bound = 1; // south of player - bounce down
-            ball.y += ballHeight;
-        } else if (ball.y < player.y - playerHeight) {
-            bound = 3; // north of player - bounce up
-            ball.y -= ballHeight;
-        }
-
-        let angle = this.getAngleFromBounds(bound);
-        ball.bounce(angle, player.velocity.x, player.velocity.y);
-    }
-
-    onCollisionGoalBall(goal, ball) {
-        let player = goal.owner;
-        global.io.sockets.in(this.gameid).emit('goalscored', {id:player.id});
-        player.lives--;
-
-
-        if(player.lives == 0 ){
-
-            goal.isActive = false;
-
-            this.killPlayer(player);
-            if(Object.keys(this.players).length === 1){
-                this.endGame();
-            }
-        }
-        else{
-            goal.setImmunity(1500);
-            this.resetBallPosition();
-        }
-    }
-
-    endGame(){
-        let winningPlayer;
-        for(let  playerKey in this.players){
-            winningPlayer = this.players[playerKey];
-        }
-        global.io.sockets.in(this.gameid).emit('endgame', {id:winningPlayer.id});
-        this.destroy();
-    }
-
-    destroy(){
-        systems.removeFromUpdater(this.updaterID);
-        delete this.collisionManager; // clean up the game
-    }
-
-    setBallVelocityBetween(ball, min, max, axisLow, axisHigh){
-        let newVelocity = {x:0,y:0};
-        while(!isVelocityBetweenMinAndMax(newVelocity, min,max)){
-            newVelocity = randomVelocity( axisLow, axisHigh);
-        }
-        ball.setVelocity(newVelocity.x, newVelocity.y)
-    }
-
-    resetBallPosition(){
-        let ball = this.balls[this.lastBallID];
-        ball.x = physObjects.gameWidth / 2;
-        ball.y = physObjects.gameHeight / 2;
-        ball.setVelocity(0,0);
-        // pause the ball for a second - we can do something clientside to make this nicer
-        setTimeout(()=>{this.setBallVelocityBetween(ball, 5, 10, -10, 10);}, 2000)
-
-    }
-
-    onCollisionBallBounds(ball, bounds) {
-        let angle = this.getAngleFromBounds(bounds);
-        ball.bounce(angle);
-    }
-
-    getAngleFromBounds(bounds) {
-        let angle;
-        switch (bounds) {
-            case 0:
-                angle = Math.PI; // east of bound - (x+)
-                break;
-            case 1:
-                angle = Math.PI / 2; // south of bound - (y+)
-                break;
-            case 2:
-                angle = 0; // west of bound - (x-)
-                break;
-            case 3:
-                angle = -Math.PI / 2; // north of bound - (y-)
-                break;
-        
-            default:
-                angle = 0; // might need better default
-                break;
-        }
-        return angle;
-    }
-
-    sendUpdateMessage(){
-
-    }
-}
-
-
-// trigger update when it starts this will recall the update
-// this should be done in the inards of a game class
-
-
-
-
-systems.startUpdate();
